Remove Uri model
Replace the Uri foreignkey by a code field directly attached to the Resource model
This makes the model simpler and solves issues with uri generation on the admin screen,
which were especially problematic when dealing with inline views.
Since Resource is abstract, this means that the codes will only not be unique across different models.